make ui functional
