+++ app/javascript/packs/stylesheets/ReactSelectStyling.js
@@ -8,13 +8,13 @@ const cursorPointerStyle = {
 };
 
 const cursorPointerStyleLg = {
-  option: base => ({ ...base, cursor: 'pointer' }),
+  option: base => ({ ...base, cursor: 'pointer', minHeight: '2.25rem' }),
   control: base => ({ ...base, fontSize: '1.25rem', fontWeight: 400, fontFamily: 'Arial', color: colors.datepickerGray, paddingLeft: '0.25rem' }),
 };
 
 const vaccineModalSelectStyling = {
   menu: base => ({ ...base, zIndex: 9999 }),
-  option: base => ({ ...base, minHeight: 30, cursor: 'pointer' })
+  option: base => ({ ...base, minHeight: 30, cursor: 'pointer' }),
 };
 
 const preferredContactTimeSelectStyling = {
@@ -30,10 +30,13 @@ const bootstrapSelectTheme = (theme, size) => {
   return {
     ...theme,
     borderRadius: 0,
-    spacing: size === 'lg' ? {
-      ...theme.spacing,
-      controlHeight: '3rem',
-    } : theme.spacing,
+    spacing:
+      size === 'lg'
+        ? {
+            ...theme.spacing,
+            controlHeight: '3rem',
+          }
+        : theme.spacing,
   };
 };
