+++ config/routes.rb
@@ -82,6 +82,7 @@ Rails.application.routes.draw do
   get '/jurisdictions/paths', to: 'jurisdictions#jurisdiction_paths', as: :jurisdiction_paths
   get '/jurisdictions/allpaths', to: 'jurisdictions#all_jurisdiction_paths', as: :all_jurisdiction_paths
   post '/jurisdictions/assigned_users', to: 'jurisdictions#assigned_users_for_viewable_patients', as: :assigned_users_for_viewable_patients
+  post '/jurisdictions/common_exposure_cohorts', to: 'jurisdictions#common_exposure_cohorts', as: :common_exposure_cohorts
 
   post '/close_contacts', to: 'close_contacts#create'
   post '/close_contacts/:id', to: 'close_contacts#update'
